requirements.txt
locustio
pyquery